fix: 태그 API로 설정 (member에서 제외)
